Add virtio-drivers (Cleanup?)
Add the virtio-drivers crate; implement a Hal for Coconut that uses
the ghcb mmio functions, and manages bounce-buffers in host-shared memory.

The buffers use SharedBox and are currently limited to 1 page in size.
The SharedBoxes need to be kept around between the share() and unshare()
calls, thus a global manager was added that holds them.

In the future, the virtio-drivers::Hal could be made stateful and the
transport could hold an Hal instance. This would allow us to keep the
shared memory management structure in there, avoiding the global.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>