kernel: add block device driver for virtio-blk
Add a block device driver that uses the virito-blk via
MMIO transport.

Requests are split into multiple of at most 4kB in size to avoid the
need of bounce buffers larger than one page. This is not a problem for
the intended use-cases and could be changed later if needed.

The driver can be enabled via the "virtio-drivers" cargo feature.

Co-developed-by: Oliver Steffen <osteffen@redhat.com>
Signed-off-by: Oliver Steffen <osteffen@redhat.com>
Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>