kernel: add in-svsm tests for VirtioBlkDriver
Add read and write tests for the VirtioBlkDriver and enable
the "virtio-drivers"" feature for the in-svsm tests.

Random data is read from/written to a disk image and the
sha256 sums calulcated in SVSM are then compared to the ones
seen on the host to test the whole path of the data.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>