kernel/fw_cfg: add find_virtio_mmio_address()
Add function to get the MMIO base address for virtio devices from
the fw_cfg file "etc/sev/svsm-virtio-mmio0".
It is expected to contain the address in hexadecimal notation
without a 0x prefix.

ToDo: Could improve the read_string() function to take an optional
max length, so it can be used in the find_file() function too.