kernel/fw_cfg: add support for virtio-mmio via hardware-info
Make use of the fw_cfg file "etc/hardware-info" to obtain the
MMIO base addresses for virtio devices.
This mechanism already exists OVMF [1] and is being added to Qemu [2].
It is intended to pass information about hardware devices to the
firmware which can not be communicated in other ways. We introduce a new
device type for SVSM for use in that file which specifies a MMIO
address. A device ID for SVSM use has been reserved in EDK2 [3].

[1] https://edk2.groups.io/g/devel/message/86051
[2]
https://lore.kernel.org/all/20250225163031.1409078-2-kraxel@redhat.com/
[3] https://github.com/tianocore/edk2/commit/f25e3d0d2c556723792bb216c030836604b4a89d

Signed-off-by: Oliver Steffen <osteffen@redhat.com>