kernel/fw_cfg: add support for virtio-mmio via hardware-info
Make use of the fw_cfg file "etc/hardware-info" to obtain the
MMIO base addresses for virtio devices.
This mechanism already exists OVMF [1] and is being added to Qemu [2].
It is intended to pass information about hardware devices to the
firmware which can not be communicated in other ways. We introduce a new
device type for SVSM for use in that file which specifies a MMIO
address.

[1] https://edk2.groups.io/g/devel/message/86051
[2]
https://lists.nongnu.org/archive/html/qemu-devel/2025-03/msg00501.html

Signed-off-by: Oliver Steffen <osteffen@redhat.com>