kernel/fw_cfg: add support for virtio-mmio via hardware-info
Make use of the fw_cfg file "etc/hardware-info" to obtain the
MMIO base addresses for virtio devices.
This mechanism was/is beeing added to OVMF [1] and Qemu [2] to
pass information about hardware devices to the firmware, which can not
be communicated in other ways. We introduce a new device type for SVSM
which specifies a MMIO address.

[1] https://edk2.groups.io/g/devel/message/86051
[2] https://lists.nongnu.org/archive/html/qemu-devel/2025-02/msg03923.html