scripts: Fix QEMU launcher nocc mode on no-cc hardware
The cbit utility does not work on non-cc hardware.
Skip the detection of the C-bit position in nocc mode.

When running in nocc mode on non-cc hardware, we can't use the kvm
backend and need to use tcg instead. To avoid another command line
option for the accelerator, let's simply always use tcg in nocc
mode.

When running with the tcg, the EPYC-v4 vcpu model does not work.
Let's use `max` with the `smep` instructions enabled instead.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>